fix(spark): avoid signed int64 overflow UB in count_min_sketch hashLong
Perform the multiply, unsigned shift, addition and mask in hashLong entirely
in uint64_t. The previous signed int64_t addition (hash += hash >> 32) is
undefined behavior on overflow in C++, whereas Java uses wrapping long
arithmetic. Output is unchanged: the PRIME_MODULUS mask keeps only the low
31 bits, which are identical regardless of shift signedness, so golden and
parity tests are unaffected.

+++ velox/functions/sparksql/aggregates/CountMinSketchAggregate.cpp
@@ -200,11 +200,14 @@ struct CountMinSketchAccumulator {
   // Hash function for integral types matching Spark's CountMinSketchImpl.hash.
   // Uses unsigned arithmetic to avoid signed overflow UB (Java wraps; C++ UB).
   int32_t hashLong(int64_t item, int32_t i) const {
-    int64_t hash = static_cast<int64_t>(
-        static_cast<uint64_t>(hashA_[i]) * static_cast<uint64_t>(item));
+    // Perform every step in uint64_t to reproduce Java's wrapping long
+    // arithmetic; signed int64_t overflow (in the multiply and the addition
+    // below) is undefined behavior in C++.
+    uint64_t hash =
+        static_cast<uint64_t>(hashA_[i]) * static_cast<uint64_t>(item);
     hash += hash >> 32;
-    hash &= kPrimeModulus;
-    return static_cast<int32_t>(hash) % width_;
+    hash &= static_cast<uint64_t>(kPrimeModulus);
+    return static_cast<int32_t>(hash % static_cast<uint64_t>(width_));
   }
 
   // Hash function for binary/string types matching Spark's
